Repro attempt 1
Repro'd an issue - though not clear if it's the same issue as Azure/azure-sdk-for-net#25648 since DAC is resolving to a VSCredential (which I needed to add after deployment to KV) - but had to disable AppConfig configuration because its challenge response was not what we typically expect and does not have a resource like KV would. AppConfig service says it supports Azure AD, but all samples with which I'm familiar with use key-based auth or connection strings.