Repro attempt 3
The problem ended up being a reused local reference variable that was captured when configuring App Configuration and Key Vault, but resolved last during runtime to the Key Vault host name. That said, I'm still unable to repro that in this attempt even when reusing both a local string reference and Uri reference.